Fix a warning

+++ c3/et.c
@@ -774,7 +774,7 @@ void CommitTransaction(void)
   revision_t newrev = -(cur_time + 1);
   assert(newrev & 1);
   ACCESS_ONCE(stm_local_revision) = newrev;
-  fprintf(stderr, "%p: stm_local_revision = %ld\n", d, newrev);
+  fprintf(stderr, "%p: stm_local_revision = %ld\n", d, (long)newrev);
   assert(d->local_revision_ref = &stm_local_revision);
 
   UpdateChainHeads(d, cur_time, localrev);
